+++ primitives/header-chain/tests/tests.rs
@@ -1,0 +1,7 @@
+mod justification {
+	mod equivocation;
+	mod optimizer;
+	mod strict;
+}
+
+mod implementation_match;
